refactor: split macro_tests.rs into tests/macro/ directory
Split the 2932-line monolithic test file into 25 files organized by
feature area (basic, interpolation, control_flow, spacing, c_each,
meta_if, join, etc.). Same 172 tests, same coverage.
