+++ tests/macro/main.rs
@@ -1,0 +1,26 @@
+mod helpers;
+
+mod args_tuple;
+mod basic;
+mod blank_lines;
+mod c_each;
+mod comments;
+mod control_flow;
+mod equivalence;
+mod escaping;
+mod if_else_chain;
+mod import_propagation;
+mod indent_dedent;
+mod integration;
+mod interpolation;
+mod join;
+mod keyword_spacing;
+mod line_splitting;
+mod meta_if;
+mod multi_language;
+mod non_brace_langs;
+mod object_literals;
+mod open_override;
+mod real_world;
+mod spacing;
+mod wrap_points;
